backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# LlamaCPP
llama-cpp-python>=0.2.56

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client (for embeddings service)
httpx>=0.26.0

# Monitoring
prometheus-client>=0.19.0

# Utilities
python-dotenv>=1.0.0
